requirements.txt
predibase>=0.1.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
jupyter>=1.0.0
pytest>=7.0.0
transformers>=4.30.0
torch>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
